intermediate commit while encode decode works
